pyproject.toml
[tool.poetry]
name = "llm-data-engineering"
version = "0.1.0"
description = ""
authors = ["frndo1203 <oliveirafer06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
rsa = "4.9"
aiohappyeyeballs = "2.4.0"
aiohttp = "3.10.5"
aiosignal = "1.3.1"
alembic = "1.13.2"
annotated-types = "0.7.0"
anyio = "4.4.0"
apache-airflow = "2.10.0"
apache-airflow-providers-common-compat = "1.2.0"
apache-airflow-providers-common-io = "1.4.0"
apache-airflow-providers-common-sql = "1.16.0"
apache-airflow-providers-fab = "1.3.0"
apache-airflow-providers-ftp = "3.11.0"
apache-airflow-providers-http = "4.13.0"
apache-airflow-providers-imap = "3.7.0"
apache-airflow-providers-postgres = "5.12.0"
apache-airflow-providers-smtp = "1.8.0"
apache-airflow-providers-sqlite = "3.9.0"
apispec = "6.6.1"
argcomplete = "3.5.0"
asgiref = "3.8.1"
attrs = "24.2.0"
babel = "2.16.0"
bardapi = "1.0.0"
blinker = "1.8.2"
browser-cookie3 = "0.19.1"
cachelib = "0.9.0"
cachetools = "5.5.0"
certifi = "2024.8.30"
cffi = "1.17.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
clickclick = "20.10.2"
colorama = "0.4.6"
colorlog = "6.8.2"
configupdater = "3.2"
connexion = "2.14.2"
cron-descriptor = "1.4.5"
croniter = "3.0.3"
cryptography = "43.0.0"
dataclasses-json = "0.6.7"
deprecated = "1.2.14"
dill = "0.3.8"
distro = "1.9.0"
dnspython = "2.6.1"
docutils = "0.21.2"
email-validator = "2.2.0"
flask = "2.2.5"
flask-appbuilder = "4.5.0"
flask-babel = "2.0.0"
flask-caching = "2.3.0"
flask-jwt-extended = "4.6.0"
flask-limiter = "3.8.0"
flask-login = "0.6.3"
flask-session = "0.5.0"
flask-sqlalchemy = "2.5.1"
flask-wtf = "1.2.1"
frozenlist = "1.4.1"
fsspec = "2024.6.1"
google-api-core = "2.19.2"
google-api-python-client = "2.143.0"
google-auth = "2.34.0"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.1"
google-re2 = "1.1.20240702"
googleapis-common-protos = "1.65.0"
greenlet = "3.0.3"
grpcio = "1.66.1"
gunicorn = "23.0.0"
h11 = "0.14.0"
h2 = "4.1.0"
hpack = "4.0.0"
httpcore = "1.0.5"
httplib2 = "0.22.0"
httpx = "0.27.2"
hyperframe = "6.0.1"
idna = "3.8"
importlib-metadata = "8.4.0"
importlib-resources = "6.4.4"
inflection = "0.5.1"
itsdangerous = "2.2.0"
jeepney = "0.8.0"
jinja2 = "3.1.4"
jiter = "0.5.0"
jmespath = "1.0.1"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
langchain = "0.2.15"
langchain-community = "0.2.15"
langchain-core = "0.2.37"
langchain-experimental = "0.0.64"
langchain-google-community = "1.0.8"
langchain-openai = "0.1.23"
langchain-text-splitters = "0.2.2"
langsmith = "0.1.108"
lazy-object-proxy = "1.10.0"
limits = "3.13.0"
linkify-it-py = "2.0.3"
lockfile = "0.12.2"
loguru = "0.7.2"
lz4 = "4.3.3"
mako = "1.3.5"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.22.0"
marshmallow-oneofschema = "3.1.1"
marshmallow-sqlalchemy = "0.28.2"
mdit-py-plugins = "0.4.1"
mdurl = "0.1.2"
methodtools = "0.4.7"
more-itertools = "10.4.0"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
oauthlib = "3.2.2"
openai = "1.43.0"
opentelemetry-api = "1.27.0"
opentelemetry-exporter-otlp = "1.27.0"
opentelemetry-exporter-otlp-proto-common = "1.27.0"
opentelemetry-exporter-otlp-proto-grpc = "1.27.0"
opentelemetry-exporter-otlp-proto-http = "1.27.0"
opentelemetry-proto = "1.27.0"
opentelemetry-sdk = "1.27.0"
opentelemetry-semantic-conventions = "0.48b0"
ordered-set = "4.1.0"
orjson = "3.10.7"
packaging = "24.1"
pandas = "2.2.2"
pathspec = "0.12.1"
pendulum = "3.0.0"
pluggy = "1.5.0"
prison = "0.2.1"
proto-plus = "1.24.0"
protobuf = "4.25.4"
psutil = "6.0.0"
psycopg2 = "2.9.9"
psycopg2-binary = "2.9.9"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pycparser = "2.22"
pycryptodomex = "3.20.0"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pygments = "2.18.0"
pyjwt = "2.9.0"
pyparsing = "3.1.4"
python-daemon = "3.0.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-gemini-api = "2.4.12"
python-nvd3 = "0.16.0"
python-slugify = "8.0.4"
pytz = "2024.1"
pyyaml = "6.0.2"
referencing = "0.35.1"
regex = "2024.7.24"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
requests-toolbelt = "1.0.0"
rfc3339-validator = "0.1.4"
rich = "13.8.0"
rich-argparse = "1.5.2"
rpds-py = "0.20.0"
setproctitle = "1.3.3"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "1.4.53"
sqlalchemy-jsonfield = "1.0.2"
sqlalchemy-utils = "0.41.2"
sqlparse = "0.5.1"
tabulate = "0.9.0"
tenacity = "8.3.0"
termcolor = "2.4.0"
text-unidecode = "1.3"
tiktoken = "0.7.0"
time-machine = "2.15.0"
tqdm = "4.66.5"
typing-inspect = "0.9.0"
typing-extensions = "4.12.2"
tzdata = "2024.1"
uc-micro-py = "1.0.3"
unicodecsv = "0.14.1"
universal-pathlib = "0.2.3"
uritemplate = "4.1.1"
urllib3 = "2.2.2"
werkzeug = "2.2.3"
wirerope = "0.4.7"
wrapt = "1.16.0"
wtforms = "3.1.2"
yarl = "1.9.6"
zipp = "3.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
